requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.7.0
transformers>=4.30.0
torch>=2.0.0
shap>=0.41.0
pyarrow>=12.0.0
pyyaml>=6.0
praw>=7.7.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
spacy>=3.5.0
networkx>=3.0
# praw, pyarrow, xgboost, shap, scikit-learn, psutil, requests, pyyaml are kept for other project steps but not needed for this analysis
# Optional: google-colab for Colab environment
# google-colab>=0.0.1a1 